requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow-cpu==2.18.0
numpy==1.26.4
gunicorn==21.2.0
huggingface_hub==0.21.4
python-dotenv==1.0.1
werkzeug==2.0.3 